Apply AppView Context is not only FrameNode
